domain connections refactored
