+++ api-js/src/domain/objects/index.js
@@ -1,2 +1,3 @@
 export * from './domain'
+export * from './domain-connection'
 export * from './domain-status'
